+++ src/Http/Requests/Admin/UpdateSettingsRequest.php
@@ -5,6 +5,7 @@ namespace Modules\Custom\MakerBid\Http\Requests\Admin;
 use Illuminate\Foundation\Http\FormRequest;
 use Modules\Custom\MakerBid\Http\Concerns\FlattensValidationErrors;
 use Modules\Custom\MakerBid\Support\BlankToNull;
+use Modules\Custom\MakerBid\Support\BidRules;
 use Modules\Custom\MakerBid\Support\BooleanishFields;
 use Modules\Custom\MakerBid\Support\SettingsRules;
 
@@ -30,9 +31,12 @@ class UpdateSettingsRequest extends FormRequest
         }
         $this->coerceBooleanFields($bools);
         $this->nullBlankFields(array_merge(
-            ['nav_label', 'nav_insert', 'default_job_status'],
+            ['nav_label', 'nav_insert', 'default_job_status', 'bid_allow'],
             array_map(static fn (string $page): string => $page.'_body', array_keys(SettingsRules::pages())),
         ));
+        if ($this->exists('bid_allow')) {
+            $this->merge(['bid_allow' => BidRules::normalizeAllow($this->input('bid_allow'))]);
+        }
     }
 
     /**